Pre Processing/requirements.txt
# Scientific Data Handling
numpy==2.4.4
pandas==3.0.2
xarray==2025.10.1
netCDF4==1.7.3
h5netcdf==1.7.3

# Geospatial and Image Processing
geopandas==1.1.1
shapely==2.1.2
rasterio==1.4.3
affine==2.4.0
scipy==1.17.1

# Utilities
python-dotenv==1.2.2